Fix mksnapshot on ppc
... which started failing after https://crrev.com/c/4672408
as addi Operand fell outside the range of uint16.

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4678894
Commit-Queue: Milad Farazmand <mfarazma@redhat.com>
Cr-Commit-Position: refs/heads/main@{#88920}

+++ src/builtins/ppc/builtins-ppc.cc
@@ -111,7 +111,9 @@ void ResetFeedbackVectorOsrUrgency(MacroAssembler* masm,
   __ LoadU8(scratch1,
             FieldMemOperand(feedback_vector, FeedbackVector::kOsrStateOffset),
             scratch2);
-  __ andi(scratch1, scratch1, Operand(~FeedbackVector::OsrUrgencyBits::kMask));
+  __ andi(
+      scratch1, scratch1,
+      Operand(static_cast<uint8_t>(~FeedbackVector::OsrUrgencyBits::kMask)));
   __ StoreU8(scratch1,
              FieldMemOperand(feedback_vector, FeedbackVector::kOsrStateOffset),
              scratch2);
